feat: implement filterDataSelector
